100DaysOfCode_Dylan/code_tracker/requirements.txt
absl-py==0.11.0
matplotlib==3.3.4
numpy==1.19.4
python-git-wrapper==0.9.0.post2
Pillow==8.1.0
requests==2.25.1
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
